Implement a workaround to artificially extend the lifetime of async DBG_VALUEs.
A better solution would be to lower async dbg.declare intrinsics into
MMI side-table entries. This patch, however, is less risky to
implement in the swift/release/5.9 branch.

rdar://107687703
